My style guide should include a labeled section outlining the colors to be used.
Each color should have:

A name.
A box displaying the colour itself.
The CSS value used to create the colour (in hex, rgb(), rgba(), hsl(), or hsla() - I chose rgb()).
An optional description of how I intend to use the colours.
I put some example colours for now, but they are not for practical use. It is to help me get started with the process of laying out the section.